Refactor locators filename
